Refactor presaved places layout in WifiActivity
The layout for the presaved places section in `activity_wifi.xml` has been updated.
The "Presaved Places" label and "+ Add" button are now contained within a horizontal `LinearLayout`.
This allows the label to be centered while the button is aligned to the right.
Margins have also been adjusted for better visual spacing.